+++ src/api/noteapi.h
@@ -45,4 +45,5 @@ class NoteApi : public QObject, public Note {
     Q_INVOKABLE QString toMarkdownHtml(bool forExport = true);
     Q_INVOKABLE QString getFileURLFromFileName(const QString &localFileName);
     Q_INVOKABLE static bool allowDifferentFileName();
+    void copy(const Note &note);
 };