+++ src/api/tagapi.cpp
@@ -47,3 +47,28 @@ void TagApi::copy(const Tag &tag)
     _parentId = tag.getParentId();
     _priority = tag.getPriority();
 }
+
+/**
+ * Returns all linked notes
+ */
+QQmlListProperty<NoteApi> TagApi::notes() {
+    _notes.clear();
+    
+    Tag tag = Tag::fetch(_id);
+    QVector<Note> notes = tag.fetchAllLinkedNotes();
+
+    QVectorIterator<Note> itr(notes);
+    while (itr.hasNext()) {
+        Note note = itr.next();
+
+        auto* noteApi = new NoteApi();
+        noteApi->copy(note);
+        _notes.append(noteApi);
+    }
+
+#if (QT_VERSION < QT_VERSION_CHECK(5, 15, 0))
+    return {this, _notes};
+#else
+    return {this, &_notes};
+#endif
+}